Remove accidental pycache files
