+++ card_dealer/__init__.py
@@ -1,13 +1,2 @@
-"""Утилиты для работы с игральными картами."""
-
-from .sorter import is_card_back, sort_by_back
-from .deck import count_cards, deduplicate, exclude_cards, find_missing_cards
-
-__all__ = [
-    "is_card_back",
-    "sort_by_back",
-    "deduplicate",
-    "exclude_cards",
-    "count_cards",
-    "find_missing_cards",
-]
+"""Обёртка для обратной совместимости."""
+from card_sorter import *  # noqa: F401,F403
