+++ card_sorter/__init__.py
@@ -1,0 +1,9 @@
+"""Пакет для распознавания и сортировки игральных карт."""
+
+__all__ = [
+    "devices",
+    "recognition",
+    "models",
+    "data",
+    "core",
+]
