+++ card_sorter/data/loader.py
@@ -1,0 +1,10 @@
+"""Загрузчик файлов данных."""
+from pathlib import Path
+import csv
+
+
+def load_cards_csv(path: Path) -> list[dict[str, str]]:
+    """Загрузить ``cards.csv`` как список словарей."""
+    with path.open(newline="") as f:
+        reader = csv.DictReader(f)
+        return [row for row in reader]
