+++ card_sorter/models/base_model.py
