+++ card_sorter/models/template_matcher.py
@@ -1,0 +1,11 @@
+"""Модуль для поиска соответствия шаблонам."""
+from __future__ import annotations
+
+from pathlib import Path
+
+
+class TemplateMatcher:
+    """Простейшая заглушка для сопоставления шаблонов."""
+
+    def match(self, image: Path) -> str:
+        return "unknown"
