+++ card_sorter/ui/controls.py
@@ -1,0 +1,7 @@
+"""Компоненты управления Streamlit."""
+
+import streamlit as st
+
+
+def show_controls() -> None:
+    st.button("Захватить карту")
