+++ card_sorter/ui/dashboard.py
@@ -1,0 +1,8 @@
+"""Заглушка для веб-интерфейса."""
+
+import streamlit as st
+
+
+def show_dashboard() -> None:
+    st.title("Card Sorter")
+    st.write("Интерфейс пока в разработке")
